requirements.txt
torch==2.0.0
torchvision==0.15.1
numpy==1.24.2
matplotlib==3.7.1
tqdm==4.65.0
pillow==9.5.0
opencv_python==4.8.1.78
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.7.0.72
scikit_image==0.19.3
pandas==2.0.3
imgaug==0.4.0
scipy==1.10.1
scikit_learn==1.3.0